+++ pokeapp/src/styling/colorschemes/_colors.scss
@@ -1,5 +1,5 @@
 $base-color: hsl(0, 0%, 11%);
-
+$secondary-color: #e3dac9;
 
 $component-container: hsl(0, 0%, 20%);
