+++ pokeapp/src/styling/components/_filterbar.scss
@@ -1,20 +1,6 @@
 @import '../colorschemes/colors';
 @import '../typography/fonts';
 
-
-.active {
-    height: 20;
-    width: 80px;
-    background-color: red;
-}
-
-.inactive {
-    height: 20;
-    width: 80px;
-    background-color: #fdfdfd;
-    border: 1px solid black;
-}
-
 .filter-bar {
     background-color: $component-container;
     grid-area: 2 / 2 / 3 / 3;
@@ -23,6 +9,7 @@
     display: flex;
     justify-content: flex-end;
     align-items: center;
+    position: relative;
     
     &-icon {
         margin: 1em; 
@@ -31,4 +18,63 @@
     &-text {
         margin: .7em;
     }
+}
+
+.filter-bar-dropdown {
+    background-color: $component-container;
+    position: absolute;
+    top: 100%;
+    right: 0;
+    z-index: 30;
+    border-radius: 10px;
+    padding: 10px;
+}
+
+.active {
+    background-color: red;
+}
+
+.inactive {
+    background-color: transparent;
+}
+
+.filter-checkbox,
+.sort-btn,
+.sort-option-btn {
+    color: $secondary-color;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    padding: 5px 10px;
+    border: 2px solid $secondary-color;
+    border-radius: 10px;
+    margin-right: 5px;
+    margin-bottom: 5px;
+    cursor: pointer;
+}
+
+.checkbox-type {
+    width: 500px;
+    display: flex;
+    justify-content: flex-end;
+    flex-wrap: wrap;
+}
+
+.sort-container {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-end;
+    margin-top: 20px;
+}
+
+.sort-type-container {
+    display: flex;
+}
+
+.sort-option-btn {
+    margin-top: 20px;
+}
+
+.sort-option-container {
+    display: flex;
 }
